search logic fixed need styling
